unittests: Add a new base test class for environment setup
This will allow all tests, even those that don't need all of the
helpers provided by the BasePlatformTest to still shared some of the
environment setup from other tests.
