+++ run_mypy.py
@@ -90,6 +90,7 @@ additional = [
     'tools',
     'docs/genrefman.py',
     'docs/refman',
+    'unittests/baseclass.py',
     'unittests/helpers.py',
 ]
